Ajout de nouvelles routes pour la santé et le webhook de playlist, ainsi que la refactorisation du serveur Express pour intégrer ces fonctionnalités.
